show earning line chart
