+++ packages/web-app/src/modules/earn-views/EarningLineChartContainer.tsx
@@ -1,6 +1,6 @@
 import { connect } from '../../connect'
 import type { RootStore } from '../../Store'
-import { EarningLineChart } from './components'
+import { EarningLineChart } from './components/EarningLineChart'
 
 const mockedEarningsPerMachine = (daysShowing: number) => {
   const shouldShowAmPm = daysShowing === 1
